많이 건드렸어요

+++ src/main/java/com/app/babybaby/controller/likeController/NowKidsLikeController.java
@@ -24,16 +24,18 @@ public class NowKidsLikeController {
     @ResponseBody
     public Boolean clickLike(Long nowKidsId, Boolean isLike){
         Boolean result = null;
-        Long sessionId = 110L;
+        Long sessionId = null;
         log.info("내가 가져온 nowKidsId는  : " + String.valueOf(nowKidsId));
         log.info("내가 가져온 isLike는  : " + String.valueOf(isLike));
 //        isLike가 false라는 뜻은 아직 좋아요가 눌리지 않은 상태
-        if(!isLike){
-            nowKidsLikeService.likeSave(nowKidsId, sessionId);
-            result = true;
-        } else{
-            nowKidsLikeService.deleteLike(nowKidsId, sessionId);
-            result = false;
+        if(sessionId != null){
+            if(!isLike){
+                nowKidsLikeService.likeSave(nowKidsId, sessionId);
+                result = true;
+            } else{
+                nowKidsLikeService.deleteLike(nowKidsId, sessionId);
+                result = false;
+            }
         }
         return result;
     }
